components/ImageProcessing/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
